requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.0
Django==6.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.0
psycopg==3.3.2
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
